move clockWidget to widgets, add namespace for widget instances, move local variable from update to class
