+++ Assets/Scripts/Widgets/Widget.cs
@@ -2,6 +2,9 @@ using UnityEngine;
 
 namespace Widgets
 {
+    /// <summary>
+    /// Abstract class from which different widget types can inherit
+    /// </summary>
     public abstract class Widget : MonoBehaviour
     {
